dje-monitor/requirements.txt
# DJE Monitor - Dependências
# Core
httpx>=0.25.0
beautifulsoup4>=4.12.0

# PDF
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2>=2.9.0

# Scheduler
APScheduler>=3.10.0

# CLI
rich>=13.0.0

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-cov>=4.1.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Importação de planilhas
openpyxl>=3.1.0
python-dateutil>=2.8.0

# Task queue
dramatiq[redis]>=1.15.0
redis>=5.0.0

# Busca Semântica
sentence-transformers>=2.7.0
qdrant-client>=1.9.0
einops>=0.7.0

# LLM
openai>=1.0.0

# Autenticação
bcrypt>=4.0.0
PyJWT>=2.8.0
email-validator>=2.1.0
